% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{AppInsightsAPI}
\alias{AppInsightsAPI}
\title{Title}
\usage{
AppInsightsAPI(RequestType = "metrics", Request)
}
\arguments{
\item{Request}{}
}
\description{
Title
}
\examples{
Direct API Call

}
