% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageAPICredentials.R
\name{CheckAPICredentials}
\alias{CheckAPICredentials}
\title{Check API Crendtials}
\usage{
CheckAPICredentials()
}
\value{
Prints API Details from Enviroment to Console.
}
\description{
Check API Crendtials
}
\examples{
CheckAPICredenrials
> CheckAPICredentials()
[1] "API KEY: "
[1] "AI HostName: "
[1] "Application ID "
}
