% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{Query}
\alias{Query}
\title{Pass to Query API}
\usage{
Query(Query)
}
\arguments{
\item{Query}{Query String used for calling API}
}
\value{
Parsed JSON
}
\description{
Used to pass queries that you construct using the Log Analytics/App Insights Query Language.
}
\examples{
Query(Query="requests\%7C\%20where\%20timestamp\%20\%3E\%3D\%20ago(24h)\%7C\%20count") # Counts Server Request from last 24hours
Query(Query=paste("requests | where timestamp >= ago(",hours,"h) | summarize percentiles(duration,50,95)") )
}
\references{
https://docs.loganalytics.io/
}
