% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageAPICredentials.R
\name{SetUpAPICredentials}
\alias{SetUpAPICredentials}
\title{Connect To API Method}
\usage{
SetUpAPICredentials(APIKEY, ApplicationID,
  Hostname = "api.applicationinsights.io")
}
\arguments{
\item{APIKEY}{Your Azure Application Insights API KEY}

\item{ApplicationID}{Your applications ID}

\item{Hostname}{The Hostname of the API. Default is api.applicationinsights.io}
}
\description{
Connect To API Method
}
\examples{
SetUpAPICredentials(APIKEY ="DEMO_KEY", Hostname ="https://api.applicationinsights.io/", ApplicationID = "DEMO_APP") # COnnect to the API Demo Application
}
